retrofit throws 400

+++ app/src/main/java/com/company/market/data/Order.kt
@@ -8,6 +8,6 @@ data class Order(
     @PrimaryKey val product_id: String,
     val product_name: String,
     val price: Int,
-    val total_item_cost: Int,
-    val quantity: Int
+    val total_item_cost: Double,
+    val quantity: Double
 )
